WebScrapingTools

Documentation for WebScrapingTools.

This package provides utilities for scraping dynamic web pages. I tried using Webdriver.jl but couldn't figure out how to configure it and get started.

Currently only Firefox with geckodriver is supported.

External Program Management

Dynamic web page scrapeing depends on having a live web browser perform any JavaScript or whatever to interpret the web page to produce the page's DOM tree.

We currently use the combination of firefox and geckodriver to process the web page. FirefoxGeckodriverSession is used to encapsulate the processes for the external firefox and geckodriver commands. The functions startup, isactive, and teardown can be used to managhe these processes.

Fetching a Web Page

The function fetch_page is used to fetch a web page. It returns a parsed HTML DOM tree if successful. Otherwise an error is thrown.

Index

WebScrapingTools.CSSSelector

WebScrapingTools.CloseWindow

WebScrapingTools.ElementClick

WebScrapingTools.FindElement

WebScrapingTools.FindElements

WebScrapingTools.FirefoxGeckodriverSession

WebScrapingTools.GetCurrentURL

WebScrapingTools.GetElementText

WebScrapingTools.GetPageSource

WebScrapingTools.GetTitle

WebScrapingTools.GetWindowHandles

WebScrapingTools.Locator

WebScrapingTools.NavigateTo

WebScrapingTools.NewSession

WebScrapingTools.WebdriverCommand

WebScrapingTools.WebdriverElement

WebScrapingTools.WebdriverSession

WebScrapingTools.WebdriverStatus

WebScrapingTools.WindowHandle

WebScrapingTools.element_text

WebScrapingTools.fetch_page

WebScrapingTools.find_element

WebScrapingTools.find_elements

WebScrapingTools.get_title

WebScrapingTools.get_window_handles

WebScrapingTools.isactive

WebScrapingTools.json_payload

WebScrapingTools.startup

WebScrapingTools.teardown

WebScrapingTools.uri_path

WebScrapingTools.with_webdriver_session

Definitions

WebScrapingTools.CSSSelector — Type

CSSSelector(css_selector_string)

https://www.w3.org/TR/webdriver2/#dfn-css-selector

WebScrapingTools.CloseWindow — Type

CloseWindow()

Closes the session's top level browsing context. https://www.w3.org/TR/webdriver2/#close-window

WebScrapingTools.ElementClick — Type

ElementClick(element_id::WebdriverElement)

Clicks on the specified element. https://www.w3.org/TR/webdriver2/#dfn-element-click

WebScrapingTools.FindElement — Type

FindElement(locator)

Webdriver command to find the first element that matches locator. https://www.w3.org/TR/webdriver2/#find-element

WebScrapingTools.FindElements — Type

FindElements(locator)

Webdriver command to find multiple elements. https://www.w3.org/TR/webdriver2/#find-elements

WebScrapingTools.FirefoxGeckodriverSession — Type

FirefoxGeckodriverSession

FirefoxGeckodriverSession encapsulates the external processes that are necessary to scrape web pages using Firefox.

WebScrapingTools.GetCurrentURL — Type

GetCurrentURL()

Webdriver command to get the current URL. https://www.w3.org/TR/webdriver2/#dfn-get-current-url.

WebScrapingTools.GetElementText — Type

GetElementText(element_id::WebdriverElement)

Returns the descendant text of the element identified by element_id. https://www.w3.org/TR/webdriver2/#dfn-get-element-text

WebScrapingTools.GetPageSource — Type

GetPageSource()

Webdriver command to get the content of the current web page. https://www.w3.org/TR/webdriver2/#dfn-get-page-source.

WebScrapingTools.GetTitle — Type

GetTitle()

Gets the page title. https://www.w3.org/TR/webdriver2/#get-title

WebScrapingTools.GetWindowHandles — Type

GetWindowHandles()

Returns the handles of the open browser windows. https://www.w3.org/TR/webdriver2/#get-window-handles

WebScrapingTools.Locator — Type

LLocators are used by various Webdriver commands that search a document for specific elements.

WebScrapingTools.NavigateTo — Type

NavigateTo(url)

Webdriver command to navidate to the specified uri. [https://www.w3.org/TR/webdriver2/#dfn-navigate-to](https://www.w3.org/TR/webdriver2/#dfn-navigate-to}.

WebScrapingTools.NewSession — Type

NewSession()

Webdriver command for creating a new session. https://www.w3.org/TR/webdriver2/#dfn-new-sessions.

WebScrapingTools.WebdriverCommand — Type

WebdriverCommand

Abstract superclass for qll supported Webdriver commands.

WebScrapingTools.WebdriverElement — Type

WebdriverElement encapsulates an element id. These are used to identify an element to Webdriver.

WebScrapingTools.WebdriverSession — Type

WebdriverSession

WebdriverSession is the abstract supertype for the types used to manage the external programs that are needed for Webdriver scraping. There would be one subtype per browser.

WebScrapingTools.WebdriverStatus — Type

WebdriverStatus()

Webdriver status command. https://www.w3.org/TR/webdriver2/#dfn-status.

WebScrapingTools.WindowHandle — Type

WindowHandle encapsulates the handle of a browser window."

WebScrapingTools.element_text — Method

element_text(session::WebdriverSession, element::WebdriverElement)

Returns the descendant text of the specified element.

See GetElementText

WebScrapingTools.fetch_page — Method

fetch_page(uri)

Fetch the dynamic content of the specified web page.

WebScrapingTools.find_element — Method

find_element(session::WebdriverSession, locator::Locator)::WebdriverElement

Returns a reference (as a WebdriverElement) to a single element. an error is thrown if no element matches locator.

Uses the FindElement command.

WebScrapingTools.find_elements — Method

find_elements(session::WebdriverSession, locator::Locator)::Vector{WebdriverElement}

Returns a list of element references (as WebdriverElements) to alll of the elements that match locator.

Uses the FindElements command.

WebScrapingTools.get_title — Method

get_title(session)

Returns the title of the curre nt page. See GetTitle

WebScrapingTools.get_window_handles — Method

get_window_handles(session::WebdriverSession)

Returns a list of handles of open browser windows.

See GetWindowHandles

WebScrapingTools.isactive — Function

isactive(::WebdriverSession)

Returns true if the session is ready to serve Webdriver commands.

WebScrapingTools.json_payload — Method

json_payload(cmd::WebdriverCommand)

Returns the payload for the HTTP request that will be sent for `cmd`.

WebScrapingTools.startup — Function

startup(::WebdriverSession)

starts the processes that are required for this type of browser.

WebScrapingTools.teardown — Function

teardown(::WebdriverSession)

terminates the processes that are required for this type of browser.

WebScrapingTools.uri_path — Function

uri_path(cmd::WebdriverCommand, ::WebdriverSession)

Returns the URI path for the request represented by cmd.

WebScrapingTools.with_webdriver_session — Method

with_webdriver_session(::Function, ::WebdriverSession)

Ensure that the WebdriverSession has running processes, then call the function.

The sessions processes are terminated once function returns.

The return value of the function is returned.