WebScrapingTools

Documentation for WebScrapingTools.

Tgis package provides utilities for scraping dynamic web pages. I tried using Webdriver.jl but couldn't figure out how to configure it and get started.

WebScrapingTools.GetCurrentURL

WebScrapingTools.GetPageSource

WebScrapingTools.NavigateTo

WebScrapingTools.NewSession

WebScrapingTools.WebdriverCommand

WebScrapingTools.WebdriverStatus

WebScrapingTools.json_payload

WebScrapingTools.uri_path

WebScrapingTools.GetCurrentURL — Type

GetCurrentURL(session_id::String)

Webdriver command to get the current URL. https://www.w3.org/TR/webdriver2/#dfn-get-current-url.

WebScrapingTools.GetPageSource — Type

GetPageSource(session_id::String)

Webdriver command to get the content of the current web page. https://www.w3.org/TR/webdriver2/#dfn-get-page-source.

WebScrapingTools.NavigateTo — Type

NavigateTo(session_id::String), url)

Webdriver command to navidate to the specified uri. [https://www.w3.org/TR/webdriver2/#dfn-navigate-to](https://www.w3.org/TR/webdriver2/#dfn-navigate-to}.

WebScrapingTools.NewSession — Type

NewSession()

Webdriver command for creating a new session. https://www.w3.org/TR/webdriver2/#dfn-new-sessions.

WebScrapingTools.WebdriverCommand — Type

WebdriverCommand

Abstract superclass for qll supported Webdriver commands.

WebScrapingTools.WebdriverStatus — Type

WebdriverStatus()

Webdriver status command. https://www.w3.org/TR/webdriver2/#dfn-status.

WebScrapingTools.json_payload — Method

json_payload(cmd::WebdriverCommand)

Returns the payload for the HTTP request that will be sent for `cmd`.

WebScrapingTools.uri_path — Function

uri_path(cmd::WebdriverCommand)

Returns the URI path for the request represented by cmd.